changed to 2 fins and corrected relays

+++ rocketArduino/7_landing.cpp
@@ -1,7 +1,7 @@
 #include "globals.h"
 
 bool landingStarted = false;
-int landingStartTime;
+long landingStartTime;
 
 void landing() {
   message += "landing, ";
